Add new declarations
This commit is mainly about add new declarations. Those declarations are
added:
* Constructor to construct a identity matrix.
* Constructor to construct with a `std::initializer_list`.
* Constructor to construct a diagnonal matrix.
* Assaignment from `std::initializer_list`.
* A method to calculate matrix exponentiation.

Because we now use `pow` to calculate matrix exponentiation, the old
`pow`s are renamed with `powNumber` and `numberPow`. The single thread
version functions are renamed with `powNumberSinglthread` and
`numPowSingleThread`. By the way, we found `*pow*` suitable for member
functions, so we adjust them as member functions.

We can use move to reduce copy, so now add a rule that all the `output`
in `*SingleThread` should not be at the same address with input
parameters.

Update some comments (the comments are for developers): make them more
readable.

We add comments (the comments are for developers and users) for the methods
in Matrix class.

Besides, we move all the definitions of Matrix and Shape out of the
classes.