Fix errors when types of matrix are different
In `clang-tidy`, we define a rule that can not compare two pointers
which have different types. When matrix `a` and matrix `output` have
different types, the `clang-tidy` will give an error (because we use
`assert(&a != &output)` to let the developers know the `a` and the
`output` should not be the same). And this error will now occur in the
`git action`, because the `git action` will build the `release` version,
the `assert` is removed by compiler. Now we use
`reinterpret_cast<const void *>` to convert them to `const void *`.
Besides, we add some simple tests for matrix whose types are different.

See #46 .