Update some declarations
We find that some declarations are wrong, so we just update them, now we
overwrite some methods in `matrix.h`, for example, there are two kinds
of `Matrix::pow` now, the first type is `Matrix::pow(number)`, the
second is `Matrix::pow(number, output)`, the second one will store the
result in `output` the first one will change `*this` directly.

We now use `matrix_declaration.h` to remove circular dependency.