Add type traits for Number
We find that when you call `a + b` and `a` and `b` are both matrices, it
will match three template functions: `operator+(Matrix, Matrix)`,
`operator+(Matrix, Number)` and `operator+(Number, Matrix)`. So we
define a type trait to check if a template is `mca::Matrix` and we use
`std::enable_if` to enable the functions whose parameters have `Number`
if and only if the `Number` is not `mca::Matrix` class.