Enhance unit tests
We found that the single-thread mode parts of multi-thread calculations are
not covered by unit tests. So we rewrite some unit tests to cover those
parts. Besides, we update some tests to make them more random.

During this process, We find that there are some multi-thread calculation
in main thread will exceed the address boundary, we use `size()` to make
sure this will not happen.

See #74.

Besides, We make `clear` of the thread poll more reasonable, now `clear`
will clear task queue, too.

In addition, we find that when generating coverage reports, the unit tests
will run in multi-thread except for the first one. This is because we
don't use `init(0)` to make sure the single-mode running in single-mode,
and use command `./mca_unit_test` to run will not run every unit test in a
new process compared with `ctest`. So we add `init(0)` in the `TearDown`
part to make sure the size of the thread pool is zero before running
unit tests.

What's more, We find that if use the default value (the number of `cpu`s),
this process will never finish on `git action`, so we use `THREAD_NUM` to
limit the number of threads, the variable is set be `10` which means
there will be `11` (`10` sub threads and `1` main thread) threads join
calculation.