Add a class called Diag
Now we define a new class called `Diag`, this class receives a container
that can be the parameter of `std::data`, e.g `std::vector`,
`std::initializer_list`. This class is for specifying a diagonal matrix.
For example, now if you want to create a diagonal matrix, you must use
`Matrix<> a(Diag({1, 2, 3}))` to create a diagonal matrix whose diagonal
elements are `1, 2, 3`.

See #90.

Besides, we find that assignments from a pointer, `std::vector` and
`std::initializer_list` are not reasonable, so we remove the
assignments. Now if you want to assign from a `std::vector`, you must
use `a = Matrix<>(vec)`, which will make the codes more readable. The
same applies to others (`a = Matrix<>(a.shape(), pointer, a.size())` and
`a = Matrix<>({{1, 2, 3}, {2, 3, 4}})`).

We also update `IdentityMatrix`. We find this class should only be
constructed from the construct with no parameters. And we remove some
redundant `explicit`, because `explicit` is for preventing implicit
conversions, there is no need to specify `explicit` for copy
constructors, move constructors, and constructors with no parameters.