Add a class called CalculationTaskNum
We find that using `std::pari` as the return type of
`threadCalculationTaskNum` is not readable. Because it is hard to
determine whether `res.first` is the thread calculation or the
`res.second`. Therefore we use `CalculationTaskNum` as the return type
now, and `res.calculation` is the calculation and `res.taskNum` is the
number of tasks, which is more readable for developers.

Solved #115.